requirements.txt
streamlit==1.28.1
groq==0.4.1
pandas==2.1.3
python-dotenv==1.0.0